Moved appearanceToOverrides to /shared, minor cleanup
